third_party/WebKit/LayoutTests/editing/execCommand/clipboard-access-with-userGesture-expected.txt

To manually test, click one of the buttons below. The result should be two 'PASS' events for either button.

Copy Cut
PASS successfullyParsed is true

TEST COMPLETE
PASS document.execCommand('copy') is true
PASS copyEventSeen is true
PASS document.execCommand('cut') is true
PASS cutEventSeen is true
